requirements-no-versions.txt
aiohappyeyeballs
aiohttp
aiosignal
alembic
annotated-types
anyio
apispec
appnope
argcomplete
argon2-cffi
argon2-cffi-bindings
arrow
asgiref
asttokens
async-lru
async-timeout
attrs
babel
backcall
beautifulsoup4
bleach
blinker
cachelib
cachetools
certifi
cffi
charset-normalizer
click
clickclick
colorama
colorlog
comm
ConfigUpdater
connexion
contourpy
copier
cron-descriptor
croniter
cryptography
cycler
db-dtypes
debugpy
decorator
defusedxml
Deprecated
dill
distlib
dnspython
docopt
docutils
dotenv
dunamai
email_validator
et_xmlfile
eval_type_backport
exceptiongroup
executing
fastjsonschema
filelock
Flask
Flask-AppBuilder
Flask-Babel
Flask-Caching
Flask-JWT-Extended
Flask-Limiter
Flask-Login
Flask-Session
Flask-SQLAlchemy
Flask-WTF
fonttools
fqdn
frozenlist
fsspec
funcy
google-api-core
google-auth
google-auth-oauthlib
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-core
google-cloud-storage
google-crc32c
google-re2
google-resumable-media
googleapis-common-protos
greenlet
grpcio
grpcio-status
gunicorn
h11
httpcore
httpx
idna
importlib_metadata
importlib_resources
inflection
ipykernel
ipython
ipython-genutils
ipywidgets
isoduration
itsdangerous
jedi
Jinja2
jinja2-ansible-filters
jmespath
joblib
json5
jsonpointer
jsonschema
jsonschema-specifications
jupyter_client
jupyter-console
jupyter_core
jupyter-events
jupyter-lsp
jupyter_server
jupyter_server_terminals
jupyterlab
jupyterlab_pygments
jupyterlab_server
jupyterlab_widgets
kiwisolver
lazy-object-proxy
limits
linkify-it-py
lockfile
lxml
Mako
markdown-it-py
MarkupSafe
marshmallow
marshmallow-oneofschema
marshmallow-sqlalchemy
matplotlib
matplotlib-inline
mdit-py-plugins
mdurl
methodtools
mistune
more-itertools
mpmath
msgspec
multidict
narwhals
nbclient
nbconvert
nbformat
nest-asyncio
networkx
notebook
notebook_shim
numpy
oauthlib
openpyxl
opentelemetry-api
opentelemetry-exporter-otlp
opentelemetry-exporter-otlp-proto-common
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-exporter-otlp-proto-http
opentelemetry-proto
opentelemetry-sdk
opentelemetry-semantic-conventions
ordered-set
overrides
packaging
pandas
pandas-gbq
pandocfilters
parso
pathspec
pendulum
pexpect
pickleshare
pillow
pip
pipenv
pipreqs
pkgutil_resolve_name
platformdirs
plotly
pluggy
plumbum
prison
prometheus_client
prompt-toolkit
propcache
proto-plus
protobuf
psutil
psycopg2
ptyprocess
pure_eval
pyarrow
pyasn1
pyasn1_modules
pycel
pycparser
pydantic
pydantic_core
pydata-google-auth
Pygments
PyJWT
pyparsing
python-daemon
python-dateutil
python-dotenv
python-json-logger
python-multipart
python-nvd3
python-pptx
python-slugify
pytz
pyxlsx
PyYAML
pyzmq
qtconsole
QtPy
questionary
referencing
requests
requests-oauthlib
requests-toolbelt
rfc3339-validator
rfc3986-validator
rich
rich-argparse
rpds-py
rsa
ruamel.yaml
ruamel.yaml.clib
scipy
seaborn
Send2Trash
setproctitle
setuptools
six
sniffio
soupsieve
SQLAlchemy
SQLAlchemy-JSONField
SQLAlchemy-Utils
sqlparse
stack-data
starlette
sympy
tabulate
tenacity
termcolor
terminado
text-unidecode
time-machine
tinycss2
tomli
torch
tornado
traitlets
types-python-dateutil
typing_extensions
tzdata
uc-micro-py
unicodecsv
universal_pathlib
uri-template
urllib3
virtualenv
wcwidth
webcolors
webencodings
websocket-client
Werkzeug
whittaker-eilers
widgetsnbextension
wirerope
wrapt
WTForms
XlsxWriter
yarg
yarl
zipp

